+++ both/views/pages/PageDashboard.jsx
@@ -7,7 +7,7 @@ PageDashboard = React.createClass({
 	render() {return(
     <div className="wrapper wrapper-content animated fadeIn">
         <div className="row">
-            <div className="col-lg-12">
+            <div className="col-lg-6 col-md-6 col-sm-12">
 	            <Box title="Work Requests">
 	            	<div className="btn-group">
 	            		<button className="btn btn-white"><i className="fa fa-th-list"></i> Day</button>
@@ -29,41 +29,34 @@ PageDashboard = React.createClass({
 	            		<button className="btn btn-white"><i className="fa fa-th"></i> 3 Month</button>
 	            	</div>
 	            	<div className="row" style={{textAlign:"center"}}>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Requests" value={{thisMonth:15,lastMonth:20}}/>
 				        </div>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Open Requests" value={{thisMonth:15,lastMonth:20}} color="green"/>
 				        </div>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Open Orders" value={{thisMonth:8,lastMonth:20}} color="orange"/>
 				        </div>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Open Quotes" value={{thisMonth:5,lastMonth:8}}/>
 				        </div>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Expired Insurance" value={{thisMonth:5,lastMonth:7}} color="green"/>
 				        </div>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Supplier Reviews" value={{thisMonth:5,lastMonth:8}} color="orange"/>
 				        </div>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Open Requests" value={{thisMonth:5,lastMonth:7}} color="green"/>
 				        </div>
-						<div className="col-lg-3 col-md-3 col-sm-6">
+						<div className="col-lg-3 col-md-3 col-sm-6" style={{padding:0}}>
 							<ProgressArc title="Open Orders" value={{thisMonth:5,lastMonth:8}} color="orange"/>
 				        </div>
 		            </div>
 	            </Box>
-	            <Box title="Preventative Maintenence Planner">
-	            	<div className="btn-group">
-	            		<button className="btn btn-white"><i className="fa fa-th-list"></i> Day</button>
-	            		<button className="btn btn-white"><i className="fa fa-th-large"></i> Week</button>
-	            		<button className="btn btn-white"><i className="fa fa-th"></i> Month</button>
-	            		<button className="btn btn-white"><i className="fa fa-th"></i> 3 Month</button>
-	            	</div>
-	            	<Calendar />
-	            </Box>
+	        </div>
+            <div className="col-lg-6 col-md-6 col-sm-12">
 	            <Box title="Active Building Compliance">
 	            	<div className="btn-group">
 	            		<button className="btn btn-white"><i className="fa fa-th-list"></i> Day</button>
@@ -73,7 +66,10 @@ PageDashboard = React.createClass({
 	            	</div>
 	            	<BarChart />
 	            </Box>
-			</div>
+	            <Box title="Preventative Maintenence Planner">
+	            	<Calendar />
+	            </Box>
+	        </div>
 		</div>
 	</div>
 	);}